refactor: split domain models into package and align with FAIR Genomes
- Replace src/domain/models.py with a src/domain/models/ package grouped
  by source domain: clinical, sequencing, wsi, radiology, sync
- Rename fields on existing domain classes to match FAIR Genomes naming
  (gender_at_birth, imaging_modality/body_region/imaging_procedure,
  SequencingRun → Sequencing, umis_present, sequencing_instrument_model,
  percentage_q30/tr20, other_quality_metrics, *_data_location, etc.)
- Delete src/application/dto/ layer; HttpCatalogueGateway now serialises
  domain objects directly via asdict()
- Update all four builders and import sites accordingly
