+++ src/application/builders/radiology_builder.py
@@ -16,9 +16,9 @@ class RadiologyBuilder:
             source_id=resolve_source_id(payload, str(payload["accession_number"])),
             imaging_study_identifier=payload.get("imaging_study_identifier"),
             belongs_to_person=payload.get("belongs_to_person"),
-            imaging_modalities=payload.get("imaging_modalities"),
-            body_regions=payload.get("body_regions"),
-            imaging_procedures=payload.get("imaging_procedures"),
+            imaging_modality=payload.get("imaging_modality"),
+            body_region=payload.get("body_region"),
+            imaging_procedure=payload.get("imaging_procedure"),
             reason_for_imaging_procedure=payload.get("reason_for_imaging_procedure"),
             study_start_date=payload.get("study_start_date"),
             dicom_series_count=payload.get("dicom_series_count"),
