--- src/application/dto/__init__.py
@@ -1,17 +1,0 @@
-from application.dto.catalogue_context_dto import (
-    build_clinical_catalogue_dto,
-    build_material_catalogue_dto,
-    build_personal_catalogue_dto,
-)
-from application.dto.catalogue_radiology_dto import build_radiology_catalogue_dto
-from application.dto.catalogue_sequencing_dto import build_sequencing_catalogue_dto
-from application.dto.catalogue_wsi_dto import build_wsi_catalogue_dto
-
-__all__ = [
-    "build_clinical_catalogue_dto",
-    "build_material_catalogue_dto",
-    "build_personal_catalogue_dto",
-    "build_radiology_catalogue_dto",
-    "build_sequencing_catalogue_dto",
-    "build_wsi_catalogue_dto",
-]
